Ensure UpdatedOnUtc is updated each time inventory changes

+++ src/Plugins/Nop.Plugin.Misc.PolyCommerce/Models/PolyCommerceCheckForShippedOrdersModel.cs
@@ -6,6 +6,6 @@ namespace Nop.Plugin.Misc.PolyCommerce.Models
 {
     public class PolyCommerceCheckForShippedOrdersModel
     {
-        public int[] OrderIds { get; set; }
+        public long[] OrderIds { get; set; }
     }
 }
